Use bytes to represent the Astarte JSON Interfaces
To represent an Astarte interface we chose to use all json
instead of major and minor name,this also improves
integration with the Astarte SDK Rust.
The AstarteMessage contains only the interface name
because it is enough to identify an Astarte interface.
Update the Readme with new proto message types.

Signed-off-by: Antonio Gisondi <antonio.gisondi@secomind.com>