feat(cli): add the -c/--config option
The option -t/--toml is unusual and less explicative than a config flag
since it indicates the format of the configuration file. The new one
should be preferred so we can free the -t flag (for timeouts for
example). Both the flags (config and toml) now require PathBuf as a
type.

Signed-off-by: Joshua Chapman <joshua.chapman@joshuachp.dev>